requirements.txt
pandas
tabula-py
jpype1
